EigenState Evolution Method for TDSE
The given Jupyter Notebook contains detailed python code (with appropriate comments) for eigenstate evolution method to solve Time Dependent Schrodinger Wave Equation. The method is applied to a model Symmetric Double Well Potential. The Method can only be applied to time independent potential functions only. As, with inclusion of time dependent potential, the entire Hamiltonian becomes time dependent and an exponential of time integral needed to be solved which is not straightaway to solve both analytically and numerically at this level {waiting for you to update the code :) } 
Hint: Break the Hamiltonian into smaller timestamps for which the Hamiltonian can be considered time independent.